requirements.txt
astroid==1.5.2
backports.functools-lru-cache==1.3
colorama==0.3.9
configparser==3.5.0
Django==1.9.10
django-annoying==0.10.3
django-braces==1.10.0
django-debug-toolbar==1.6
django-easy-pdf==0.1.0
django-endless-pagination==2.0
django-generic-json-views==0.8
django-jenkins==0.110.0
django-mathfilters==0.4.0
django-seed==0.1.6
django-wkhtmltopdf==3.1.0
docopt==0.6.2
enum34==1.1.6
fake-factory==9999.9.9
Faker==0.7.7
html5lib==1.0b8
ipaddress==1.0.17
isort==4.2.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
MySQL-python==1.2.5
numpy==1.12.0
pep8==1.7.0
Pillow==3.4.2
pipreqs==0.4.6
pylint==1.7.1
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.3.0
requests==2.13.0
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.2
virtualenv==15.1.0
webencodings==0.5
wrapt==1.10.10
xhtml2pdf==0.0.6
xlwt==1.2.0
yarg==0.1.9
XlsxWriter==0.9.9
dbfread==2.0.7